Added Person Entity
Added Person Entity to Core Data and setting up 2 people named Jane and Bob. There is also a fetch request in the application function that outputs to the console to confirm that they are in the app.

This commit builds fully with not bugs or issues.

Note: Still Nothing is being saved to core data at the time of this commit. The data is completely wiped every time the app is shut down.