Added subclassed to the Core Data Model
## item 1:
In this commit there have been **2 subclasses created from the Core Data Model**. Each subclass represents one of the attributes of the Core Data Model.

4 New classes auto created. These 2 are the ones that changes should be made to...
Person.swift
Device.Swift

## item 2:
All view controllers and the app delegate where updated to no longer initialize as NSManagedObject, instead they **initialize as Device type** and this allows for simplification of much of the fetching and setting code.

Note: this code is currently building completely and is safe for a regression point if needed.