Add functionality to the Person tab
This commit adds functionallity to the People tab.

The tab now allow the user to click on a persons name and see a list of the devices that are assosiated with them using the relationships that were setup in the previous commit.

This commit is fully building and bug free.